requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-crispy-forms==1.9.0
django-heroku==0.3.1
djangorestframework==3.11.0
get==2019.4.13
gunicorn==20.0.4
idna==2.9
Pillow==7.1.2
post==2019.4.13
psycopg2==2.8.5
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.0.1
